Add add-on for Primo VE for removing GES request
This add-on will remove the configured GES request {{displayLabel}} from all the items in Full view. The request will still be available on Locations and Libraries that approved by {{libraryCode}} and {{subLocationCode}}. The places can single place or array with commas.

We added an option to the description that can filter the system (primo or VE) by the "systemExclusive" field.